ADD menu for calendar
