+++ Net14/Calendar/CalendarDrawer.cs
@@ -32,10 +32,6 @@ namespace Calendar
                 Console.WriteLine();
             }
 
-            Console.ForegroundColor = ConsoleColor.Green;
-            Console.WriteLine("use arrowLeft or arrowRight to scrolling monthes \nUse \"Space\" for enter date.");
-            Console.ForegroundColor = ConsoleColor.White;
-
             Console.WriteLine($"{monthLevel.WeekendsCount} " +
 $"weekends and {DateTime.DaysInMonth(monthLevel.Year, monthLevel.MonthNumber) - monthLevel.WeekendsCount} working days ");
